cmake: set global vars only when this is the global project
This greatly increases compatibility when this template is used as a
subproject. Several edge cases and bugs are also fixed with this new
logic, such as version update when root v0 and local v0. Local to v1
would then also update root to v1 even if it was a subproject. New
logic uses *_SOURCE_DIR to decide if we are the root project.

Also some minor polish of the main CMakeLists.txt, moving some code to
local cmake/* files.