cicd: package c++ runtime for mingw_x86_64
The runtime used for cross compilation is not the default on Windows,
unlike with GNU/Linux and macOS. This means we have to package the
cross-compilation runtime together with the binary for C++ functionality
to work.

As the default threading model is Win32 we package for that, meaning
libwpinthread does not need to be packaged. For details see:
http://wiki.qt.io/MinGW-64-bit#GCC_Threading_model_.28posix_vs_win32.29