sphinx: set doctree cache to build tmp folder
Recent Sphinx versions, such as 1.8.0, have a bug where they put the
cache folder in SOURCE/.doctrees instead of DEST/.doctrees. We now
specify a specific tmpdir for the cache, fixing it on our end.

This change also is an optimisation for documentation packaging, since
in previous versions the .doctrees dir ended up being packaged into the
final job artifact for the sphinx_html job.